Moves all TodoList functionality to TodoList.js
